+++ CHANGELOG.md
@@ -16,6 +16,8 @@ contents to a new release section (e.g. _[0.x]_).
 
 ### Fixed
 
+- Fixed expressions with rational exponents being collected as if they had no exponents. (PLUT-1129)
+
 ### Changed
 
 ### Removed